# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BasisGuard Evidence Log & Adaptation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List treatment profiles
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, deprecated, opt_in_only]
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreatmentProfile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a treatment profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentProfile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a treatment profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile (e.g. deprecate it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreatmentProfile"
  /profiles/{id}/delta:
    get:
      operationId: getProfileDelta
      tags: [profiles]
      summary: Get delta report — positions that would change under this profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delta report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDelta"

  # ── Export ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Authority Citation ─────────────────────────────────────────────────────
...
    ProfileRule:
      type: object
      properties:
        event_type:
          type: string
        classification:
          type: string
        tier:
          type: string
          enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
        citation_ids:
          type: array
          items:
            type: string
      required:
        - event_type
        - classification
        - tier
        - citation_ids
    TreatmentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, deprecated, opt_in_only]
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ProfileRule"
        changelog:
          type: ["string", "null"]
        created_at:
          type: string
      required:
        - id
        - name
        - status
        - rules
        - created_at
    ProfileInput:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, deprecated, opt_in_only]
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ProfileRule"
        changelog:
          type: string
      required:
        - name
        - status
        - rules
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, deprecated, opt_in_only]
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ProfileRule"
        changelog:
          type: string
    ProfileDelta:
      type: object
      properties:
        profile_id:
          type: string
        profile_name:
          type: string
        changed_positions:
          type: array
          items:
            $ref: "#/components/schemas/DeltaEntry"
        total_changed:
          type: integer
      required:
        - profile_id
        - profile_name
        - changed_positions
        - total_changed
    DeltaEntry:
      type: object
      properties:
        position_id:
          type: string
        tx_id:
          type: string
        event_type:
          type: string
        before_classification:
          type: string
        after_classification:
          type: string
        before_tier:
          type: string
        after_tier:
          type: string
      required:
        - position_id
        - tx_id
        - event_type
        - before_classification
        - after_classification
        - before_tier
        - after_tier

    # ── Position Record ────────────────────────────────────────────────────────
...